No Kernel ajuste para excluir do banco se deletou os arquivos do Storage. Em PreRegistroAjaxAdminRequest ajuste para recuperar a id do pre-registro. Em PreRegistroService inserido no orderby quando status aprovado ao listar os pre-registros; dispara mensagem no log quando anexa e exclui um arquivo no formulario; e corrigido no método de excluir arquivo. Em web ajustada a variavel na rota de updateAjax para retornar a id do pre-registro na validação. E testes de log em AnexoTest.

+++ app/Console/Kernel.php
@@ -183,8 +183,9 @@ class Kernel extends ConsoleKernel
             foreach($prs as $pr)
             {
                 $paths = $pr->anexos->pluck('path')->all();
-                $pr->anexos()->delete();
-                Storage::delete($paths);
+                $deleted = Storage::delete($paths);
+                if($deleted)
+                    $pr->anexos()->delete();
             }
         })->weeklyOn(7, '3:00');
     }
